Improve the error handling for notification (#56)
Onesignal includes different types of errors in a single key. According to the documents, its type could be either map or array.
"All included players are not subscribed" is the most common error we faced in our service. There is a function "IsErrAllPlayersNotSubscribed" added to address that error. The log level of message "All included players are not subscribed" is changed the to WARN and do not submit to sentry.